@@ -10,7 +10,7 @@ namespace NQueue.Internal.Db.InMemory
 
         public ValueTask<IWorkItemDbQuery> Get() => new ValueTask<IWorkItemDbQuery>(_query);
 
-        public ValueTask EnqueueWorkItem(DbTransaction tran, TimeZoneInfo tz, Uri url, string queueName, string debugInfo,
+        public ValueTask EnqueueWorkItem(DbTransaction tran, TimeZoneInfo tz, Uri url, string? queueName, string? debugInfo,
             bool duplicateProtection)
         {
             throw new Exception("The in-memory NQueue implementation is not compatible with DB transactions.");
